First Stage. Data entry

+++ src/main/kotlin/domain/entity/Crew.kt
@@ -5,5 +5,5 @@ data class Crew(
     val name: String,
     val countOfWorkers: Int,
     val hourlyRate: Int,
-    val workIndex: Int
+    val workIndex: Double
 )
